Now, feature to update any teacher is available.

+++ F2022SchoolDb/Models/Teacher.cs
@@ -13,6 +13,6 @@ namespace F2022SchoolDb.Models
         public string teacherlname;
         public string employeenumber;
         public string hiredate;
-        public int salary;
+        public decimal salary;
     }
 }
